Fix typed predicate comparison to nil
Before this patch, the builder would compile a comparison to nil,
although since `value` in the `compare` function was not non-nil, it was
bridged to Optional<Any>, which surprisingly would not evaluate to nil
when building the ComparisonPredicate, which was represented in the
predicate string as "<null>" instead of "nil" or "NULL", causing a crash
when the predicate was evaluated.

Making the value optional here allows the builder to correctly unwrap
optionals and represent them in the expression string.